EKF: Changed get_control_mode to coyp 32 bits

+++ EKF/estimator_interface.h
@@ -253,7 +253,7 @@ public:
 	virtual void get_gyro_bias(float bias[3]) = 0;
 
 	// get EKF mode status
-	void get_control_mode(uint16_t *val)
+	void get_control_mode(uint32_t *val)
 	{
 		*val = _control_status.value;
 	}
